+++ src/index.mjs
@@ -114,7 +114,6 @@ import Video from './nodes/video.mjs'
 import Wbr from './nodes/wbr.mjs'
 import Custom from './nodes/custom.mjs'
 
-import { render } from './render.mjs'
 import { createState } from './state/state.mjs'
 
 import Router from './router/router.mjs'
@@ -122,7 +121,6 @@ import Component from './component/component.mjs'
 import I18n from './intl/i18n.mjs'
 
 export {
-  render,
   createState,
   Router,
   Component,